Added translations for readerById page.
It works, we can see reader details and we can delete it [will use yesno dialog.]

+++ src/ui/routes/textSet/readers/byId/readerById.ts
@@ -7,6 +7,9 @@ import {ReaderService} from "../../../../../services/reader/readerService";
 import {ReaderIncludeUser} from "../../../../../services/reader/reader";
 import {LoadingScreen} from "../../../../components/base/screens/loading/loadingScreen";
 import {AppNavigator} from "../../../../appNavigator";
+import {DialogManager} from "../../../../base/dialog/dialogManager";
+import {TextSet} from "../../../../../services/textSet/textSet";
+import {TextSetService} from "../../../../../services/textSet/textSetService";
 
 type ReaderByIdAttrs = { readerId: number, textSetId: number };
 
@@ -14,6 +17,7 @@ export const ReaderById = new class extends BaseComponent<ReaderByIdAttrs, any> 
     private reader: ReaderIncludeUser | undefined = undefined;
     private isError: boolean = false;
     private isDeleted: boolean = false;
+    private textSet: TextSet | undefined = undefined;
 
     override oninit(vnode: Mithril.Vnode<ReaderByIdAttrs, Mithril._NoLifecycle<any>>): any {
         ReaderService.getReaders(vnode.attrs.textSetId)
@@ -33,6 +37,11 @@ export const ReaderById = new class extends BaseComponent<ReaderByIdAttrs, any> 
                 this.isError = true;
                 redraw()
             });
+
+        TextSetService.getById(vnode.attrs.textSetId)
+            .then((textSet) => {this.textSet = textSet; redraw()})
+            .catch(() => {this.isError = true; redraw()});
+
         return super.oninit(vnode);
     }
 
@@ -47,12 +56,18 @@ export const ReaderById = new class extends BaseComponent<ReaderByIdAttrs, any> 
     private deleteReader() {
         if (!this.reader) return;
         let {textSetId, userId, id} = this.reader!!.reader;
-        ReaderService.deleteReader(textSetId, id)
-            .then(() => {
-                this.isDeleted = true;
-                redraw();
-            })
-            .catch(() => alert("Couldn't delete this reader")) // TODO
+
+        DialogManager.yesNoAsync(t("all.confirmation"), t("all.reader.really-want-to-delete-reader", {textSet: this.textSet?.title, reader: this.reader?.userName}))
+            .then((yesNo) => {
+                if (yesNo == "yes") {
+                    ReaderService.deleteReader(textSetId, id)
+                        .then(() => {
+                            this.isDeleted = true;
+                            redraw();
+                        })
+                        .catch(() => DialogManager.info(t("all.something-went-wrong"), t("reader.could-not-delete")))
+                }
+            });
     }
 
     private renderReaderData(): Vnode<any, any> {
@@ -60,30 +75,36 @@ export const ReaderById = new class extends BaseComponent<ReaderByIdAttrs, any> 
             return Layout.center(
                 Layout.block(
                     m("div",
-                        m("h1", this.reader?.reader.id!! + " został usunięty") /* TODO */,
+                        m("h1", t("all.reader.reader-is-deleted", {reader: this.reader?.userName!!})),
                         m("button.btn.btn-primary", {onclick: () => AppNavigator.textSetById(this.reader?.reader.textSetId!!)}, t("all.home"))
                     )
                 )
             )
         }
 
-        return Layout.centerNodes(
-            Layout.withHeader(t("routes.reader_by_id.title"), t("routes.reader_by_id.body") ?? undefined, m("div")),
+        if (this.reader && this.textSet) {
+            return Layout.centerNodes(
+                Layout.withHeader(t("reader-by-id.title"), t("reader-by-id.body", {textSet: this.textSet.title}) ?? undefined, m("div")),
 
-            Layout.block(
-                m("#app_reader_by_id__header",
-                    m("h2", this.reader?.userName),
-                    m("p", this.reader?.reader.id)
-                )
-            ),
+                Layout.block(
+                    m("#app_reader_by_id__header",
+                        m("h2", this.reader?.userName),
+                        m("p", this.reader?.reader.id)
+                    )
+                ),
 
-            Layout.block(
-                Layout.withHeader(t("all.text-owner-actions"), undefined,
-                    m("#app_reader_by_id__owner_actions",
-                        m("button.btn.btn-danger", {onclick: () => this.deleteReader()}, t("all.reader.delete-reader"))
+                Layout.block(
+                    Layout.withHeader(t("all.text-owner-actions"), undefined,
+                        m("#app_reader_by_id__owner_actions",
+                            m("button.btn.btn-danger", {onclick: () => this.deleteReader()}, t("all.reader.delete-reader"))
+                        )
                     )
                 )
             )
-        )
+        }
+
+        else {
+            return m("h2", t("all.something-went-wrong"))
+        }
     }
 }
